Split up type checker to multiple, specialized compile units
