+++ src/typechecker/TypeCheckerTopLevelDefinitions.cpp
@@ -1,0 +1,77 @@
+// Copyright (c) 2021-2025 ChilliBits. All rights reserved.
+
+#include "TypeChecker.h"
+
+namespace spice::compiler {
+
+std::any TypeChecker::visitMainFctDef(MainFctDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitMainFctDefPrepare(node);
+  else
+    return visitMainFctDefCheck(node);
+}
+
+std::any TypeChecker::visitFctDef(FctDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitFctDefPrepare(node);
+  else
+    return visitFctDefCheck(node);
+}
+
+std::any TypeChecker::visitProcDef(ProcDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitProcDefPrepare(node);
+  else
+    return visitProcDefCheck(node);
+}
+
+std::any TypeChecker::visitStructDef(StructDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitStructDefPrepare(node);
+  else
+    return visitStructDefCheck(node);
+}
+
+std::any TypeChecker::visitInterfaceDef(InterfaceDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitInterfaceDefPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitEnumDef(EnumDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitEnumDefPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitGenericTypeDef(GenericTypeDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitGenericTypeDefPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitAliasDef(AliasDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitAliasDefPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitGlobalVarDef(GlobalVarDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitGlobalVarDefPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitExtDecl(ExtDeclNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitExtDeclPrepare(node);
+  return nullptr;
+}
+
+std::any TypeChecker::visitImportDef(ImportDefNode *node) {
+  if (typeCheckerMode == TC_MODE_PRE)
+    return visitImportDefPrepare(node);
+  return nullptr;
+}
+
+} // namespace spice::compiler
